+++ src/teamcomm/net/logging/LogYamlLoader.java
@@ -191,7 +191,7 @@ class LogYamlLoader extends Constructor {
                         queue.addLast(new LogReplayTask.LoggedObject(time - baseTime + recordedTime, data.gameState, message));
                     } else if (recording) { // Ending the current section?
                         // end section signal
-                        queue.addLast(new LogReplayTask.LoggedObject(time - baseTime + recordedTime, -1, 14383421));
+                        queue.addLast(new LogReplayTask.LoggedObject(time - baseTime + recordedTime, (byte)-1, 14383421));
                         recordedTime += time - baseTime;
                         recording = false;
                     }
